Create BubbleSort.java
An implementation of the Bubble sort method for a list of comparable objects. A Random object is used to generate integers for the test list of integers. The number of items in the list as well as the max number to be generated for the test list can altered as desired in the field variables.